+++ TC02_Final_Project/HPS/dwht.h
@@ -17,9 +17,9 @@
  * @param M The number of columns in the matrices.  
  * @return A result of the subtraction of matrixB from matrixA (N x M).
  */
-float* diff(float *matrixA, float* matrixB, int N, int M);
+int* diff(int *matrixA, int* matrixB, int N, int M);
 
-float* multiply_matrices(float* matrixA, float* matrixB, int N, int M, int K);
+int* multiply_matrices(int* matrixA, int* matrixB, int N, int M, int K);
 
 
 /**
@@ -35,7 +35,7 @@ float* multiply_matrices(float* matrixA, float* matrixB, int N, int M, int K);
  * @note The input array size must be a power of 2 for the transform to work correctly.
  *       Ensure that the input data is properly allocated and initialized before calling this function.
  */
-float* dwht_1d(float* vec, int N);
+int* dwht_1d(int* vec, int N);
 
 /**
  * @brief Performs a 1-dimensional Fast Walsh-Hadamard Transform (FWHT) on the input data.
@@ -50,8 +50,8 @@ float* dwht_1d(float* vec, int N);
  * @note The input array size must be a power of 2 for the transform to work correctly.
  *       Ensure that the input data is properly allocated and initialized before calling this function.
  */
-float* fwht_1d(float* vec, int N);
-float* dwht_1d_inverse(float* vec, int N);
+int* fwht_1d(int* vec, int N);
+int* dwht_1d_inverse(int* vec, int N);
 
 /**
  *
@@ -65,7 +65,7 @@ float* dwht_1d_inverse(float* vec, int N);
  * @note The function modifies the input matrix directly. Ensure that the matrix is properly allocated and contains valid data before calling this function.
  * @note The dimensions N and M should be powers of 2 for optimal DWHT performance.
  */
-float* dwht_2d_octave(float* matrix, int N, int M);
+int* dwht_2d_octave(int* matrix, int N, int M);
 
 /**
  *
@@ -79,7 +79,7 @@ float* dwht_2d_octave(float* matrix, int N, int M);
  * @note The function modifies the input matrix directly. Ensure that the matrix is properly allocated and contains valid data before calling this function.
  * @note The dimensions N and M should be powers of 2 for optimal DWHT performance.
  */
-float* dwht_2d_inverse_octave(float* matrix, int N, int M);
+int* dwht_2d_inverse_octave(int* matrix, int N, int M);
 
 /**
  *
@@ -93,7 +93,7 @@ float* dwht_2d_inverse_octave(float* matrix, int N, int M);
  * @note The function modifies the input matrix directly. Ensure that the matrix is properly allocated and contains valid data before calling this function.
  * @note The dimensions N and M should be powers of 2 for optimal DWHT performance.
  */
-float* dwht_2d_octave_ll(float* matrix, int N, int M);
+int* dwht_2d_octave_ll(int* matrix, int N, int M);
 
 
 /**
@@ -108,7 +108,7 @@ float* dwht_2d_octave_ll(float* matrix, int N, int M);
  * @note The function modifies the input matrix directly. Ensure that the matrix is properly allocated and contains valid data before calling this function.
  * @note The dimensions N and M should be powers of 2 for optimal DWHT performance.
  */
-float* dwht_2d_inverse_octave_ll(float* matrix, int N, int M);
+int* dwht_2d_inverse_octave_ll(int* matrix, int N, int M);
 
 /**
  * @brief This function multiplies two matrices, matrixA and matrixB, and returns the resulting matrix.
@@ -121,7 +121,7 @@ float* dwht_2d_inverse_octave_ll(float* matrix, int N, int M);
  *
  * @return A pointer to the resulting matrix (N x K), or NULL if an error occurred.  The caller is responsible for freeing the allocated memory.
  */
-float* multiply_matrices(float* matrixA, float* matrixB, int N, int M, int K);
+int* multiply_matrices(int* matrixA, int* matrixB, int N, int M, int K);
 
 /**
  * @brief Transposes a given matrix.
@@ -134,4 +134,4 @@ float* multiply_matrices(float* matrixA, float* matrixB, int N, int M, int K);
  * @param M The number of columns in the input matrix.
  * @return A pointer to the transposed matrix (M x N). Returns NULL if memory allocation fails.
  */
-float* transpose(float *matrix, int N, int M);
+int* transpose(int *matrix, int N, int M);
